bugfix: failed to start container when /etc/mtab is symbol link
In some images, `/etc/mtab` is symbol link file, such as:
`/etc/mtab -> /proc/mounts`, when use `/bin/cp -f` to overwrite
`/etc/mtab`, it will cause error, because it will link to `/proc/mounts`
file on the host.

So we use the option of `--remove-destination` to remove the symbol file
before overwrite it.

Signed-off-by: Rudy Zhang <rudyflyzhang@gmail.com>